created TryGetTiff
